+++ src/app/shared/utils/angular-components.utils.ts
@@ -1,5 +1,0 @@
-export class AngularComponentsUtils {
-  static asset(path: string) {
-    return new URL(path, document.baseURI).toString();
-  }
-}